More assertion helpers to Runner
